+++ Xenity_Engine/Source/tools/math.h
@@ -5,7 +5,7 @@ class Vector3;
 class Math 
 {
 public:
-	static void MultiplyMatrix(const double* A, const double* B, double* result, int rA, int cA, int rB, int cB);
+	static void MultiplyMatrix(const float* A, const float* B, float* result, int rA, int cA, int rB, int cB);
 	static Vector3 GetDirectionFromAngles(const float angleA, const float angleB);
 	static Vector3 GetDirectionFromAngle(const float angleA);
 };
